+++ lib/generators/dynamo_db/session_store_migration/templates/dynamo_db_session_store.rb
@@ -1,4 +1,4 @@
-class <%= name.camelize %> < ActiveRecord::Migration[<%= migration_version %>]
+class <%= name.camelize %> < ActiveRecord::Migration<%= migration_version %>
   def up
     Aws::SessionStore::DynamoDB::Table.create_table
   end