+++ spec/action_dispatch/session/dynamodb_store_spec.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative '../../spec_helper'
+
+module ActionDispatch
+  module Session
+    describe DynamodbStore do
+      it 'does something' do
+        store = DynamodbStore.new()
+        require 'byebug'
+        byebug
+      end
+    end
+  end
+end